Added: TODO reminders, Renamed: list-feature-point to panel-feature-point
