+++ src/components/label-panel/panel/panel-feature-point.vue
