+++ src/components/menu/model/modal-select-savetype.vue
@@ -141,6 +141,7 @@ export default {
      * @param {String} message
      */
     displaySnackbar(message) {
+      // TODO: Implement a snackbar
       this.snackbarMsg = message;
     },
     /**
